+++ adapter-1.11/src/main/java/me/googas/chat/adapters/v1_11/LegacyPlayerTitleAdapter.java
@@ -2,7 +2,7 @@ package me.googas.chat.adapters.v1_11;
 
 import java.util.logging.Level;
 import lombok.NonNull;
-import me.googas.chat.ErrorHandler;
+import me.googas.chat.debug.Debugger;
 import me.googas.chat.adapters.PlayerTitleAdapter;
 import me.googas.chat.exceptions.PacketHandlingException;
 import me.googas.chat.packet.PacketType;
@@ -42,7 +42,7 @@ public final class LegacyPlayerTitleAdapter implements PlayerTitleAdapter {
           .setField(4, fadeOut)
           .send(player);
     } catch (PacketHandlingException e) {
-      ErrorHandler.getInstance()
+      Debugger.getInstance()
           .handle(Level.SEVERE, "Could not send title, subtitle and times packet", e);
     }
   }