+++ adapter/src/main/java/me/googas/chat/debug/DebugTask.java
@@ -1,0 +1,7 @@
+package me.googas.chat.debug;
+
+
+public interface DebugTask {
+
+
+}