+++ CHANGELOG.md
@@ -9,5 +9,6 @@ Changelog
   - [X] Soft dependencies classes
 - Added Javadoc for
   - [X] Custom tab API
-- [X] Renamed `Line` to `Text` 
+- [X] Renamed `Line` to `Text`
+- [X] Renamed `ErrorHandler` to `Debugger`
 - Moved `Versions#getObjectiveAdapter` to `Players#getObjectiveAdapter`